Replace row index in all template row elements
References to the row index were only replace in input elements and
certain other elements. With custom templates (possibly using the
standard widget wrapper) many more elements may contain references to
the row index. Replace the row index in all id, name, for, href and
data-fieldname attributes when cloning the template row.

To simplify the reindexing the current row index is stored in a data
attribute.