Set widget mode on cell widget.
In order to support ``plone.autoform.directives.mode`` on the cell, the
mode is set on the cell widget when updating the widgets.
This makes it possible to hide a column using the ``mode`` directive.

The behavior of ``<th>`` and ``<td>`` tags regarding "hidden"-mode was
differing: while ``<th>``s were just omitted, ``<td>``s were rendered
with a "datagridwidget-hidden-data" CSS class, hiding the cell in CSS.
For beeing consistent, the ``<th>`` behavior is changed so that the tag
is kept but the same CSS class is used.